Remove comments folder, rename count file and add movie-counter.test.js file
